refactor: Separate package file and offset
- FileRepository getting the file inputstream.
- OffsetRepository should get file offset about the file key.
