+++ kafka-streams-connect/build.gradle.kts
@@ -1,7 +1,7 @@
 plugins {
 	id("org.springframework.boot") version "3.5.6"
 	id("io.spring.dependency-management") version "1.1.7"
-	id("com.google.cloud.tools.jib") version "3.4.6"
+	id("com.google.cloud.tools.jib") version "3.4.5"
 	kotlin("plugin.spring") version "2.1.20"
 	kotlin("jvm") version "2.1.20"
 }
