+++ source-connector/src/main/java/sourceconnector/repository/file/FileRepository.java
@@ -1,10 +1,10 @@
-package sourceconnector.repository;
+package sourceconnector.repository.file;
 
 import java.io.IOException;
 import java.io.InputStream;
 
 /**
- * Get the file contents from a file remoteoffsetmanager.repository (e.g. LocalFile System, S3, GCS, Azure Blob Storage, etc.)
+ * Get the file contents from a file storage (e.g. LocalFile System, S3, GCS, Azure Blob Storage, etc.)
  */
 public interface FileRepository {
   /**
