+++ source-connector/src/main/java/sourceconnector/repository/offset/S3OffsetRecordRepository.java
@@ -1,4 +1,4 @@
-package sourceconnector.repository;
+package sourceconnector.repository.offset;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
