+++ source-connector/src/test/java/sourceconnector/repository/LocalOffsetRecordRepositoryTest.java
@@ -19,6 +19,7 @@ import org.apache.kafka.common.serialization.StringSerializer;
 import org.junit.jupiter.api.*;
 import org.springframework.kafka.config.TopicBuilder;
 import sourceconnector.domain.offset.LocalFileOffsetRecord;
+import sourceconnector.repository.offset.LocalOffsetRecordRepository;
 
 import java.util.Collections;
 import java.util.List;
